TS-3145: add additional textBuffer APIs
To make textBuffer more useful for formatting wawlls of text
  - fix textBuffer::enlargeBuffer so that it works with an initial size of 0
  - add textBuffer::format to append printf-formatted text
  - add textBuffer::chomp to remove any trailing newline
  - add textBuffer::release to remove the internal formatting buffer

+++ lib/records/P_RecCore.cc
@@ -1001,7 +1001,7 @@ int RecWriteConfigFile(textBuffer *tb)
       break;
     }
 
-    if (nbytes != tb->spaceUsed()) {
+    if (nbytes != (int)tb->spaceUsed()) {
       RecLog(DL_Warning, "write to file: %s fail, disk maybe full", tmp_filename);
       result = REC_ERR_FAIL;
       break;
